refactor: extract shared test harness into grpc-task-test-helpers.ts
