+++ packages/plugin-core/src/grpc-list-tasks.test.ts
@@ -128,70 +128,10 @@ vi.mock("./compute-task-status.js", () => ({
 
 // ── Import AFTER mocks ──────────────────────────────────────────
 
-import { registerGrackleRoutes } from "./grpc-service.js";
-import { sqlite as _sqlite, workspaceStore, taskStore, sessionStore } from "@grackle-ai/database";
-const sqlite = _sqlite!;
+import { workspaceStore, taskStore, sessionStore } from "@grackle-ai/database";
 import { computeTaskStatus } from "./compute-task-status.js";
 import { grackle } from "@grackle-ai/common";
-import type { ConnectRouter } from "@connectrpc/connect";
-
-/** Apply schema DDL to in-memory database. */
-function applySchema(): void {
-  sqlite.exec(`
-    CREATE TABLE IF NOT EXISTS workspaces (
-      id            TEXT PRIMARY KEY,
-      name          TEXT NOT NULL,
-      description   TEXT NOT NULL DEFAULT '',
-      repo_url      TEXT NOT NULL DEFAULT '',
-      environment_id TEXT NOT NULL DEFAULT '',
-      status        TEXT NOT NULL DEFAULT 'active',
-      use_worktrees INTEGER NOT NULL DEFAULT 1,
-      working_directory TEXT NOT NULL DEFAULT '',
-      default_persona_id TEXT NOT NULL DEFAULT '',
-      token_budget  INTEGER NOT NULL DEFAULT 0,
-      cost_budget_millicents INTEGER NOT NULL DEFAULT 0,
-      created_at    TEXT NOT NULL DEFAULT (datetime('now')),
-      updated_at    TEXT NOT NULL DEFAULT (datetime('now'))
-    );
-
-    CREATE TABLE IF NOT EXISTS tasks (
-      id            TEXT PRIMARY KEY,
-      workspace_id  TEXT REFERENCES workspaces(id),
-      title         TEXT NOT NULL,
-      description   TEXT NOT NULL DEFAULT '',
-      status        TEXT NOT NULL DEFAULT 'not_started',
-      branch        TEXT NOT NULL DEFAULT '',
-      depends_on    TEXT NOT NULL DEFAULT '[]',
-      assigned_at   TEXT,
-      started_at    TEXT,
-      completed_at  TEXT,
-      review_notes  TEXT NOT NULL DEFAULT '',
-      created_at    TEXT NOT NULL DEFAULT (datetime('now')),
-      updated_at    TEXT NOT NULL DEFAULT (datetime('now')),
-      sort_order    INTEGER NOT NULL DEFAULT 0,
-      parent_task_id TEXT NOT NULL DEFAULT '',
-      depth         INTEGER NOT NULL DEFAULT 0,
-      can_decompose INTEGER NOT NULL DEFAULT 0,
-      default_persona_id TEXT NOT NULL DEFAULT '',
-      workpad   TEXT NOT NULL DEFAULT '',
-      schedule_id TEXT NOT NULL DEFAULT '',
-      token_budget  INTEGER NOT NULL DEFAULT 0,
-      cost_budget_millicents INTEGER NOT NULL DEFAULT 0
-    );
-  `);
-}
-
-/** Extract service handlers via fake router. */
-function getHandlers(): Record<string, (...args: unknown[]) => unknown> {
-  let handlers: Record<string, (...args: unknown[]) => unknown> = {};
-  const fakeRouter = {
-    service(_def: unknown, impl: Record<string, (...args: unknown[]) => unknown>) {
-      handlers = { ...handlers, ...impl };
-    },
-  } as unknown as ConnectRouter;
-  registerGrackleRoutes(fakeRouter);
-  return handlers;
-}
+import { resetSchema, getHandlers } from "./grpc-task-test-helpers.js";
 
 describe("gRPC listTasks handler", () => {
   let handlers: Record<string, (...args: unknown[]) => unknown>;
@@ -199,9 +139,7 @@ describe("gRPC listTasks handler", () => {
 
   beforeEach(() => {
     vi.clearAllMocks();
-    sqlite.exec("DROP TABLE IF EXISTS tasks");
-    sqlite.exec("DROP TABLE IF EXISTS workspaces");
-    applySchema();
+    resetSchema();
 
     // Seed workspace and tasks using real stores
     workspaceStore.createWorkspace(WORKSPACE_ID, "Test Project", "desc", "", "");
