+++ packages/plugin-core/src/grpc-task-test-helpers.ts
@@ -1,0 +1,77 @@
+/**
+ * Shared in-memory database harness for gRPC task handler integration tests.
+ * Provides `resetSchema` and `getHandlers` for use by test files that already
+ * declare the required `vi.mock()` blocks (vitest hoists mocks at compile time
+ * and they cannot be re-exported from a helper module).
+ */
+import { sqlite as _sqlite } from "@grackle-ai/database";
+import { registerGrackleRoutes } from "./grpc-service.js";
+import type { ConnectRouter } from "@connectrpc/connect";
+
+const sqlite = _sqlite!;
+
+/**
+ * Drop and recreate the tasks/workspaces tables in the in-memory test database.
+ * Call this in `beforeEach` to start each test with a clean slate.
+ */
+export function resetSchema(): void {
+  sqlite.exec("DROP TABLE IF EXISTS tasks");
+  sqlite.exec("DROP TABLE IF EXISTS workspaces");
+  sqlite.exec(`
+    CREATE TABLE IF NOT EXISTS workspaces (
+      id            TEXT PRIMARY KEY,
+      name          TEXT NOT NULL,
+      description   TEXT NOT NULL DEFAULT '',
+      repo_url      TEXT NOT NULL DEFAULT '',
+      environment_id TEXT NOT NULL DEFAULT '',
+      status        TEXT NOT NULL DEFAULT 'active',
+      use_worktrees INTEGER NOT NULL DEFAULT 1,
+      working_directory TEXT NOT NULL DEFAULT '',
+      default_persona_id TEXT NOT NULL DEFAULT '',
+      token_budget  INTEGER NOT NULL DEFAULT 0,
+      cost_budget_millicents INTEGER NOT NULL DEFAULT 0,
+      created_at    TEXT NOT NULL DEFAULT (datetime('now')),
+      updated_at    TEXT NOT NULL DEFAULT (datetime('now'))
+    );
+
+    CREATE TABLE IF NOT EXISTS tasks (
+      id            TEXT PRIMARY KEY,
+      workspace_id  TEXT REFERENCES workspaces(id),
+      title         TEXT NOT NULL,
+      description   TEXT NOT NULL DEFAULT '',
+      status        TEXT NOT NULL DEFAULT 'not_started',
+      branch        TEXT NOT NULL DEFAULT '',
+      depends_on    TEXT NOT NULL DEFAULT '[]',
+      assigned_at   TEXT,
+      started_at    TEXT,
+      completed_at  TEXT,
+      review_notes  TEXT NOT NULL DEFAULT '',
+      created_at    TEXT NOT NULL DEFAULT (datetime('now')),
+      updated_at    TEXT NOT NULL DEFAULT (datetime('now')),
+      sort_order    INTEGER NOT NULL DEFAULT 0,
+      parent_task_id TEXT NOT NULL DEFAULT '',
+      depth         INTEGER NOT NULL DEFAULT 0,
+      can_decompose INTEGER NOT NULL DEFAULT 0,
+      default_persona_id TEXT NOT NULL DEFAULT '',
+      workpad   TEXT NOT NULL DEFAULT '',
+      schedule_id TEXT NOT NULL DEFAULT '',
+      token_budget  INTEGER NOT NULL DEFAULT 0,
+      cost_budget_millicents INTEGER NOT NULL DEFAULT 0
+    );
+  `);
+}
+
+/**
+ * Extract all service handlers from the gRPC router by registering routes
+ * against a fake ConnectRouter that captures the handler map.
+ */
+export function getHandlers(): Record<string, (...args: unknown[]) => unknown> {
+  let handlers: Record<string, (...args: unknown[]) => unknown> = {};
+  const fakeRouter = {
+    service(_def: unknown, impl: Record<string, (...args: unknown[]) => unknown>) {
+      handlers = { ...handlers, ...impl };
+    },
+  } as unknown as ConnectRouter;
+  registerGrackleRoutes(fakeRouter);
+  return handlers;
+}
